Move Temporality enum opentelemetry_sdk::metrics::data:: to opentelemetry_sdk::metrics::
