+++ opentelemetry-sdk/src/metrics/mod.rs
@@ -69,6 +69,29 @@ use std::hash::{Hash, Hasher};
 
 use opentelemetry::{Key, KeyValue, Value};
 
+/// Defines the window that an aggregation was calculated over.
+#[derive(Debug, Copy, Clone, Default, PartialEq, Eq, Hash)]
+#[non_exhaustive]
+pub enum Temporality {
+    /// A measurement interval that continues to expand forward in time from a
+    /// starting point.
+    ///
+    /// New measurements are added to all previous measurements since a start time.
+    #[default]
+    Cumulative,
+
+    /// A measurement interval that resets each cycle.
+    ///
+    /// Measurements from one cycle are recorded independently, measurements from
+    /// other cycles do not affect them.
+    Delta,
+
+    /// Configures Synchronous Counter and Histogram instruments to use
+    /// Delta aggregation temporality, which allows them to shed memory
+    /// following a cardinality explosion, thus use less memory.
+    LowMemory,
+}
+
 /// A unique set of attributes that can be used as instrument identifiers.
 ///
 /// This must implement [Hash], [PartialEq], and [Eq] so it may be used as
@@ -132,7 +155,7 @@ impl Hash for AttributeSet {
 mod tests {
     use self::data::{DataPoint, HistogramDataPoint, ScopeMetrics};
     use super::*;
-    use crate::metrics::data::{ResourceMetrics, Temporality};
+    use crate::metrics::data::ResourceMetrics;
     use crate::testing::metrics::InMemoryMetricExporterBuilder;
     use crate::{runtime, testing::metrics::InMemoryMetricExporter};
     use opentelemetry::metrics::{Counter, Meter, UpDownCounter};
