+++ opentelemetry-sdk/src/testing/metrics/in_memory_exporter.rs
@@ -1,8 +1,9 @@
 use crate::metrics::data;
-use crate::metrics::data::{Histogram, Metric, ResourceMetrics, ScopeMetrics, Temporality};
+use crate::metrics::data::{Histogram, Metric, ResourceMetrics, ScopeMetrics};
 use crate::metrics::exporter::PushMetricExporter;
 use crate::metrics::MetricError;
 use crate::metrics::MetricResult;
+use crate::metrics::Temporality;
 use async_trait::async_trait;
 use std::collections::VecDeque;
 use std::fmt;
